refactor: the same method for cutting edges across isochrone builders
